Refactor - moved ntosAttribute to its own modulr
